refactor: valueobject を completion/ に分割、全 __init__.py をカラに
- valueobject/chat.py を分割：
  - completion/message.py: MessageDTO
  - completion/gender.py: Gender, GenderType
  - completion/riddle.py: RiddleEvaluation, RiddleResponse

- すべての __init__.py をカラに変更し、明示的なインポートを強制
  - service/__init__.py
  - service/completion/__init__.py
  - usecase/__init__.py
  - usecase/completion/__init__.py
  - valueobject/__init__.py
  - valueobject/completion/__init__.py

- 全ファイルのインポート文を更新：
  - service/completion/ 配下の全ファイル
  - usecase/completion/ 配下の全ファイル
  - repository/chat.py
  - models.py
  - views.py
  - tests.py

既存のテストがすべてパス (12 tests OK)
