+++ llm_chat/domain/repository/chat.py
@@ -1,6 +1,6 @@
 from django.contrib.auth.models import User
 
-from llm_chat.domain.valueobject.chat import MessageDTO
+from llm_chat.domain.valueobject.completion.message import MessageDTO
 from llm_chat.models import ChatLogs
