--- llm_chat/domain/service/completion/__init__.py
@@ -1,33 +1,0 @@
-from llm_chat.domain.service.completion.base import BaseChatService
-from llm_chat.domain.service.completion.common import (
-    ChatService,
-    OpenAIChatStreamingService,
-    get_chat_history,
-)
-from llm_chat.domain.service.completion.riddle import (
-    RiddleTask,
-    RIDDLE_END_MESSAGE,
-    get_prompt,
-    create_initial_prompt,
-)
-from llm_chat.domain.service.completion.multimedia import (
-    OpenAIDalleChatService,
-    OpenAITextToSpeechChatService,
-    OpenAISpeechToTextChatService,
-)
-from llm_chat.domain.service.completion.rag import OpenAIRagChatService
-
-__all__ = [
-    "BaseChatService",
-    "ChatService",
-    "OpenAIChatStreamingService",
-    "get_chat_history",
-    "RiddleTask",
-    "RIDDLE_END_MESSAGE",
-    "get_prompt",
-    "create_initial_prompt",
-    "OpenAIDalleChatService",
-    "OpenAITextToSpeechChatService",
-    "OpenAISpeechToTextChatService",
-    "OpenAIRagChatService",
-]
