+++ llm_chat/domain/service/completion/rag.py
@@ -7,7 +7,7 @@ from lib.llm.valueobject.config import OpenAIGptConfig, ModelName
 from lib.llm.valueobject.rag import PdfDataloader
 from llm_chat.domain.repository.chat import ChatLogRepository
 from llm_chat.domain.service.completion.base import BaseChatService
-from llm_chat.domain.valueobject.chat import MessageDTO
+from llm_chat.domain.valueobject.completion.message import MessageDTO
 
 
 class OpenAIRagChatService(BaseChatService):
