+++ llm_chat/domain/service/completion/riddle.py
@@ -6,9 +6,9 @@ from lib.llm.service.completion import LlmCompletionService, BaseLLMTask
 from lib.llm.valueobject.completion import RoleType, Message
 from lib.llm.valueobject.config import OpenAIGptConfig, GeminiConfig
 from llm_chat.domain.repository.chat import ChatLogRepository
-from llm_chat.domain.valueobject.chat import (
-    MessageDTO,
-    Gender,
+from llm_chat.domain.valueobject.completion.message import MessageDTO
+from llm_chat.domain.valueobject.completion.gender import Gender
+from llm_chat.domain.valueobject.completion.riddle import (
     RiddleResponse,
     RiddleEvaluation,
 )
