+++ llm_chat/domain/usecase/completion/common.py
@@ -7,7 +7,7 @@ from lib.llm.valueobject.completion import RoleType, StreamResponse
 from lib.llm.valueobject.config import OpenAIGptConfig, GeminiConfig
 from llm_chat.domain.service.completion.common import ChatService, OpenAIChatStreamingService
 from llm_chat.domain.usecase.completion.base import UseCase
-from llm_chat.domain.valueobject.chat import MessageDTO
+from llm_chat.domain.valueobject.completion.message import MessageDTO
 
 
 class LlmChatUseCase(UseCase):
