+++ llm_chat/views.py
@@ -11,17 +11,12 @@ from dotenv import load_dotenv
 
 from lib.llm.valueobject.completion import StreamResponse
 from llm_chat.domain.repository.chat import ChatLogRepository
-from llm_chat.domain.service import RIDDLE_END_MESSAGE
-from llm_chat.domain.usecase import (
-    UseCase,
-    LlmChatUseCase,
-    RiddleUseCase,
-    OpenAIGptStreamingUseCase,
-    OpenAIDalleUseCase,
-    OpenAITextToSpeechUseCase,
-    OpenAISpeechToTextUseCase,
-    OpenAIRagUseCase,
-)
+from llm_chat.domain.service.completion.riddle import RIDDLE_END_MESSAGE
+from llm_chat.domain.usecase.completion.base import UseCase
+from llm_chat.domain.usecase.completion.common import LlmChatUseCase, OpenAIGptStreamingUseCase
+from llm_chat.domain.usecase.completion.multimedia import OpenAIDalleUseCase, OpenAITextToSpeechUseCase, OpenAISpeechToTextUseCase
+from llm_chat.domain.usecase.completion.rag import OpenAIRagUseCase
+from llm_chat.domain.usecase.completion.riddle import RiddleUseCase
 from lib.llm.valueobject.config import OpenAIGptConfig, GeminiConfig, ModelName
 from llm_chat.forms import UserTextForm
 from llm_chat.models import ChatLogs
